lara/control: port Lara_WaterCurrent to TRX
