+++ src/libtrx/include/libtrx/game/lara/control.h
@@ -5,4 +5,5 @@
 extern void Lara_InitialiseMeshes(const GF_LEVEL *level);
 
 void Lara_BaddieCollision(ITEM *lara_item, COLL_INFO *coll);
+void Lara_WaterCurrent(COLL_INFO *coll);
 void Lara_DismountVehicle(void);
