+++ src/tr2/game/lara/misc.c
@@ -159,93 +159,3 @@ void Lara_GetJointAbsPosition_I(
     vec->z = item->pos.z + (g_MatrixPtr->_23 >> W2V_SHIFT);
     Matrix_Pop();
 }
-
-void Lara_WaterCurrent(COLL_INFO *const coll)
-{
-    ITEM *const item = g_LaraItem;
-
-    int16_t room_num = g_LaraItem->room_num;
-    const ROOM *const room = Room_Get(g_LaraItem->room_num);
-    const int32_t z_sector = (g_LaraItem->pos.z - room->pos.z) >> WALL_SHIFT;
-    const int32_t x_sector = (g_LaraItem->pos.x - room->pos.x) >> WALL_SHIFT;
-    g_LaraItem->box_num =
-        Room_GetWorldSector(room, g_LaraItem->pos.x, g_LaraItem->pos.z)->box;
-
-    XYZ_32 target;
-    if (Box_CalculateTarget(&target, item, &g_Lara.lot) == TARGET_NONE) {
-        return;
-    }
-
-    target.x -= item->pos.x;
-    if (target.x > g_Lara.current_active) {
-        item->pos.x += g_Lara.current_active;
-    } else if (target.x < -g_Lara.current_active) {
-        item->pos.x -= g_Lara.current_active;
-    } else {
-        item->pos.x += target.x;
-    }
-
-    target.z -= item->pos.z;
-    if (target.z > g_Lara.current_active) {
-        item->pos.z += g_Lara.current_active;
-    } else if (target.z < -g_Lara.current_active) {
-        item->pos.z -= g_Lara.current_active;
-    } else {
-        item->pos.z += target.z;
-    }
-
-    target.y = target.y - item->pos.y;
-    if (target.y > g_Lara.current_active) {
-        item->pos.y += g_Lara.current_active;
-    } else if (target.y < -g_Lara.current_active) {
-        item->pos.y -= g_Lara.current_active;
-    } else {
-        item->pos.y += target.y;
-    }
-
-    g_Lara.current_active = 0;
-    coll->facing =
-        Math_Atan(item->pos.z - coll->old.z, item->pos.x - coll->old.x);
-    Collide_GetCollisionInfo(
-        coll, item->pos.x, item->pos.y + LARA_HEIGHT_UW / 2, item->pos.z,
-        room_num, LARA_HEIGHT_UW);
-
-    switch (coll->coll_type) {
-    case COLL_FRONT:
-        if (item->rot.x > 35 * DEG_1) {
-            item->rot.x = item->rot.x + LARA_UW_WALL_DEFLECT;
-        } else if (item->rot.x < -35 * DEG_1) {
-            item->rot.x = item->rot.x - LARA_UW_WALL_DEFLECT;
-        } else {
-            item->fall_speed = 0;
-        }
-        break;
-
-    case COLL_TOP:
-        item->rot.x -= LARA_UW_WALL_DEFLECT;
-        break;
-
-    case COLL_TOP_FRONT:
-        item->fall_speed = 0;
-        break;
-
-    case COLL_LEFT:
-        item->rot.y += 910;
-        break;
-
-    case COLL_RIGHT:
-        item->rot.y -= 910;
-        break;
-
-    default:
-        break;
-    }
-
-    if (coll->side_mid.floor < 0) {
-        item->pos.y += coll->side_mid.floor;
-        item->rot.x += LARA_UW_WALL_DEFLECT;
-    }
-    Lara_Col_Shift(coll);
-
-    coll->old = item->pos;
-}
